+++ packages/react/src/HeadContext.ts
@@ -1,0 +1,7 @@
+import { createContext } from 'react'
+
+// @ts-expect-error TS(2554): Expected 1 arguments, but got 0.
+const headContext = createContext()
+headContext.displayName = 'WreatheHeadContext'
+
+export default headContext