+++ packages/react/src/PageContext.ts
@@ -1,0 +1,7 @@
+import { createContext } from 'react'
+
+// @ts-expect-error TS(2554): Expected 1 arguments, but got 0.
+const pageContext = createContext()
+pageContext.displayName = 'WreathePageContext'
+
+export default pageContext