+++ packages/react/src/server.ts
@@ -1,3 +1,1 @@
-export const runServer = (str: string) => {
-  return console.log(str)
-}
+export { default as default } from '@wreathe-js/core/server'