+++ packages/react/CHANGELOG.md
@@ -2,9 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](https://keepachangelog.com/), and this
-project adheres to [Semantic Versioning](https://semver.org/).
-
 ## Unreleased
 
-- Initial release
+- initial release based on [@inertiajs/react](https://github.com/inertiajs/inertia) (v1.0.0-beta.2)
+
+### Refactoring
+
+- convert adapter to TypeScript
+- change namespace `Inertia/inertia` to `Wreathe/wreathe`
+- remove `microbundle` dependency
+
+### Other
+
+- add `unbuild` dependency
+- add `@types/lodash.isequal` dependency